Use 2 sources files for the C writable implementations
