+++ src/unix/lwt_unix_unix.h
@@ -45,15 +45,7 @@
    | Test for readability/writability                                |
    +-----------------------------------------------------------------+ */
 
-CAMLprim value lwt_unix_writable(value fd)
-{
-    struct pollfd pollfd;
-    pollfd.fd = Int_val(fd);
-    pollfd.events = POLLOUT;
-    pollfd.revents = 0;
-    if (poll(&pollfd, 1, 0) < 0) uerror("writable", Nothing);
-    return (Val_bool(pollfd.revents & POLLOUT));
-}
+
 
 /* +-----------------------------------------------------------------+
    | Memory mapped files                                             |
