+++ src/unix/windows_c/windows_writable.c
@@ -1,0 +1,5 @@
+#if defined(LWT_ON_WINDOWS)
+#include "lwt_unix.h"
+LWT_NOT_AVAILABLE1(unix_writable)
+#endif
+
